+++ docs/src/examples/components/Menu/Variations/MenuExampleVertical.shorthand.tsx
@@ -2,8 +2,21 @@ import * as React from 'react'
 import { Menu } from '@stardust-ui/react'
 
 const items = [
-  { key: 'editorials', content: 'Editorials' },
-  { key: 'review', content: 'Reviews' },
+  {
+    key: 'editorials',
+    content: 'Editorials',
+    icon: {
+      name: 'bookmark',
+      outline: true,
+    },
+  },
+  {
+    key: 'review',
+    content: 'Reviews',
+    icon: {
+      name: 'word',
+    },
+  },
   { key: 'events', content: 'Upcoming Events' },
 ]